changed size_t to int count

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -2,15 +2,14 @@
 #include <stdio.h>
 
 /**
- * print_list - prints all the elements of a
- * list_t list
+ * print_list - prints all the elements of a list_t list
  * @h: pointer to the head of the list
  *
  * Return: the number of nodes in the list
  */
 size_t print_list(const list_t *h)
 {
-	size_t count = 0;
+	int count = 0;
 
 	while (h != NULL)
 	{
